Fix _inverse_update to handle leaf nodes only
_inverse_update() was configured to handle all nodes with updaters, even
if those updaters were on branch nodes. This behavior is incorrect
because when `store_schema` is specified to an `Engine` constructor
before the full hierarchy is known (e.g. when an EngineProcess tunnel
has yet to be populated by the outer simulation), an updater may be set
on what will become a branch node because that node doesn't yet have any
children.